Refactor: Move SoundcloudHelper
